Change way of saving the waveform
Add option to save as separate .wav files.
But adapt to what the SoundCAM model expects: a .npy file of shape (RIRs, mic_channels, audio_samples).
Now this is set to (N, 1, 48,000), with the farthest mic and 1 second of RIR. This will be long enough, as VGGish uses a precutoff value which is lower.